poiskino/requirements.txt
asgiref==3.3.1
asn1crypto==1.4.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.2.1
defusedxml==0.7.0rc1
Django==3.1.4
django-crispy-forms==1.10.0
django-filter==2.4.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
drf-yasg==1.20.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Naked==0.1.31
numpy==1.19.4
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.8
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.9
pycryptodomex==3.9.9
PyJWT==1.7.1
pyOpenSSL==20.0.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
shellescape==3.8.1
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
